Arquitectura-Publica-Suscribe/requirements.txt
pika==0.10.0
pyfiglet==0.7.5
progressbar2==3.34.2
